fix: improve MenuGroup spacing variables
